TY_PROJECT-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
python-nmap==0.7.1
pdfplumber==0.10.3
zipfile36==0.1.3
python-magic==0.4.27
hashlib2==1.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
